args: Move named args parsing to separate class
